Don't use monospace font for property descriptions
Currently, object and array property descriptions are in regular
font, but others are forced to monospace with an inline style.

In my opinion, the monospace font looks ugly here, since descriptions
are generally human-readable text, not code, so this PR removes
the inline style.

However, if this is a desirable feature, let me know, and I will
work on an alternative approach where all descriptions are styled
consistently, and can be over-ridden in some way by the user.