Use animation rather than transition to avoid arbritrary max-height
For requests and repsonses with a lot of properties, the schema and
examples are currently truncated, because the containing elements have
max-height set to 5000px or 10000px.

This value is part of the expand/collapse transition - animating from
zero to 'auto' turns out to be a hard problem, as discussed in this
Stack Overflow question: https://stackoverflow.com/q/3508605/157957

This patch takes an alternative approach: using "animation" rather than
"transition", the last frame of the animation can have a max-height,
but then revert to auto. The "snap" to full height will happen off-screen,
unless your screen is more than 5000px tall.

This also allows a slightly more natural collapse transition, quickly
collapsing long blocks, so that shorter blocks don't have an awkward
pause before they start animating.