Make expand and collapse animations feel more responsive
Firstly, this uses "large viewport height" units to stop/start the animation
at the actual maximum visible area, rather than having to over-estimate a
pixel size.

Secondly, it immediately starts fading out the content on collapse, so there
is some immediate feedback to the user when the user clicks the element.